backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
spacy==3.8.4
gunicorn==23.0.0
python-dotenv==1.0.1 